Moved loading code into transpiler.py and updated main.py to use Transpiler class
